requirements.txt
# Core Django
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-parler==2.3
drf-yasg==1.21.7
django-filter==23.5

# Database
psycopg2-binary==2.9.10

# Image Processing
Pillow==10.2.0

# Authentication
social-auth-app-django==5.4.0
social-auth-core==4.5.2
python-jose==3.3.0
python-dotenv==1.0.1

# Utilities
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
urllib3==2.2.1
